User Guide

Overview

BitcoinRPC.jl provides a Julia wrapper to Bitcoin's JSON-RPC API. Most of the blockchain and mining RPCs have been implemented in the form of callable functions. Depending on the operation, a function can return data (either as a DataFrame or Dict) or the associated HTTP response (Dict).

Available functions

BitcoinRPC.collect_block_stats

BitcoinRPC.collect_network_stats

BitcoinRPC.show_best_block_hash

BitcoinRPC.show_block

BitcoinRPC.show_block_count

BitcoinRPC.show_block_hash

BitcoinRPC.show_block_header

BitcoinRPC.show_block_stats

BitcoinRPC.show_blockchain_info

BitcoinRPC.show_chain_tips

BitcoinRPC.show_chain_txstats

BitcoinRPC.show_difficulty

BitcoinRPC.show_mempool_ancestors

BitcoinRPC.show_mempool_descendents

BitcoinRPC.show_mempool_entry

BitcoinRPC.show_mempool_info

BitcoinRPC.show_mempool_raw

BitcoinRPC.show_mining_info

BitcoinRPC.show_network_hashps

BitcoinRPC.show_tx_out

BitcoinRPC.show_tx_out_setinfo

BitcoinRPC.verify_chain

Blockchain RPCs

BitcoinRPC.show_best_block_hash — Method

show_best_block_hash(auth::UserAuth)

Returns the hash of the best (tip) block in the most-work fully-validated chain.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Example

julia> show_best_block_hash(auth)
"00000000000000000004202bf6a4b09fcfde643a4a1d206f18949308bc505011"

BitcoinRPC.show_block — Method

show_block(auth::UserAuth; blockhash::String, verbosity::Int64 = 1)

Get block data depending on selected verbosity.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Optional keywords

verbosity::Int64 : If verbosity is 0, returns a string that is serialized, hex-encoded data for block hash. If verbosity is 1, returns an object with information about block hash. If verbosity is 2, returns an object with information about block hash and information about each transaction.

Example

julia> show_block(auth, blockhash = show_best_block_hash(auth), verbosity = 1)
Dict{String, Any} with 18 entries:
  "time"              => DateTime("2022-05-07T01:13:07")
  "difficulty"        => 2.97944e13
  "bits"              => "17097275"
  "previousblockhash" => "00000000000000000007c2e512cf0b3f99cec4334c3499f4d477babc63a1b77b"
  "merkleroot"        => "d0fda1281cec9719be434acb24a81715f98ac0948a445f1b6b4e967d8781615c"

BitcoinRPC.show_blockchain_info — Method

show_blockchain_info(auth::UserAuth)

Returns an object containing various state info regarding blockchain processing.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Example

julia> show_blockchain_info(auth)
Dict{String, Any} with 13 entries:
  "verificationprogress" => 0.999998
  "difficulty"           => 2.97944e13
  "chain"                => "main"
  "initialblockdownload" => false
  "size_on_disk"         => 459765806919
...

BitcoinRPC.show_block_count — Method

show_block_count(auth::UserAuth)

Returns the height of the most-work fully-validated chain.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Example

julia> show_block_count(auth)
735241

BitcoinRPC.show_block_hash — Method

show_block_hash(auth::UserAuth; height::Int64 = 0)

Returns hash of block in best-block-chain at height provided.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

height::Int64 : Height of the desired block, default is set to 0.

Example

julia> show_block_hash(auth, height = 696969)
"0000000000000000000b3e2716da675f99df43134a67fc1987c5590f1c370472"

BitcoinRPC.show_block_header — Method

show_block_header(auth::UserAuth; blockhash::String, verbose::Bool=true)

Returns the contents of the block header.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

blockhash::String : Hash of the desired block

Optional keywords

verbose::Bool : If verbose is false, returns a string that is serialized, hex-encoded data for blockheader hash. If verbose is true (default), returns an object with information about blockheader hash.

Result

julia> show_block_header(auth, blockhash = show_block_hash(auth, height = 700_000))
Dict{String, Any} with 15 entries:
  "time"              => DateTime("2021-09-11T04:14:32")
  "difficulty"        => 1.84152e13
  "bits"              => "170f48e4"
  "previousblockhash" => "0000000000000000000aa3ce000eb559f4143be419108134e0ce71042fc636eb"
  "nextblockhash"     => "00000000000000000002f39baabb00ffeb47dbdb425d5077baa62c47482b7e92"
...

BitcoinRPC.show_block_stats — Method

show_block_stats(auth::UserAuth; hashORheight::StringOrInt = 0, stats = "")

Compute per block statistics for a given window. All amounts are in satoshis.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Optional keywords

hashORheight::StringOrInt : Block hash or height

stats : Specific group of stats, e.g. ["avgfee", "utxo_increase"]

Example

julia> show_block_stats(auth, hashORheight = 500_000)
Dict{String, Any} with 29 entries:
  "avgtxsize"           => 388
  "time"                => DateTime("2017-12-18T18:35:25")
  "totalfee"            => 3.39352
  "utxo_increase"       => 1899
  "total_out"           => 1401737618054
...

BitcoinRPC.show_chain_tips — Method

show_chain_tips(auth::UserAuth)

Return information about all known tips in the block tree, including the main chain as well as orphaned branches.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Example

julia> show_chain_tips(auth)
1-element Vector{Any}:
 Dict{String, Any}("height" => 735239, "branchlen" => 0, "status" => "active", "hash" => ...

BitcoinRPC.show_chain_txstats — Method

show_chain_txstats(auth::UserAuth; nblocks::Int64 = 144, blockhash::String = "")

Compute statistics about the total number and rate of transactions in the chain for a given search window.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Optional keywords

nblocks::Int64 : Size of the window in number of blocks, default is 144 (~ 1 day). If blockhash is not specified, stats are calculated until latest height.

blockhash::String : The hash of the block that ends the window. Default is empty, which means stats are shown until latest height.

Example

julia-repl julia> show_chain_txstats(auth, nblocks = 1440, blockhash = show_block_hash(auth, height = 700_000)) Dict{String, Any} with 8 entries: "txcount" => 669566382 "window_tx_count" => 2512428 "time" => DateTime("2021-09-11T04:14:32") "window_interval" => 821264 "window_final_block_height" => 700000 "window_final_block_hash" => "0000000000000000000590fc0f3eba193a278534220b2b37e9849e1a770ca959" "window_block_count" => 1440 "txrate" => 3.05922`

BitcoinRPC.show_difficulty — Method

show_difficulty(auth::UserAuth)

Returns the proof-of-work difficulty as a multiple of the minimum difficulty.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Example

julia> show_difficulty(auth)
2.979440758931208e13

BitcoinRPC.show_mempool_ancestors — Method

show_mempool_ancestors(auth::UserAuth; txid::String, verbose::Bool = true)

If txid is in the mempool, returns all in-mempool ancestors.

BitcoinRPC.show_mempool_descendents — Method

show_mempool_descendents(auth::UserAuth; txid::String, verbose::Bool = true)

If txid is in the mempool, returns all in-mempool descendants.

BitcoinRPC.show_mempool_entry — Method

show_mempool_entry(auth::UserAuth; txid::String)

Returns mempool data for a given transaction.

BitcoinRPC.show_mempool_info — Method

show_mempool_info(auth::UserAuth)

Returns details on the active state of the TX memory pool.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Example

```julia-repl julia> showmempoolinfo(auth) Dict{String, Any} with 9 entries: "unbroadcastcount" => 0 "maxmempool" => 1000000000 "bytes" => 1272046 "loaded" => true "usage" => 6281104 ...

BitcoinRPC.show_mempool_raw — Method

show_mempool_raw(auth::UserAuth; verbose::Bool = false, mempool_sequence::Bool = false)

Returns all transaction ids in memory pool as a json array of string transaction ids.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Example

julia> show_mempool_raw(auth)
2225-element Vector{String}:
 "46db70c88170e584b6787adc7561a6aa06ddce51f8bcd34bd4576ce0114e55ea"
 "5e17a3b3bf10e137050734ac884fb496e5933ebc7401764f62c60b18f58df891"
 "3fa233c8bcb8f2002b4997507c36a887c76359719c643a7e93541a7230f84b9c"
 ...

BitcoinRPC.show_tx_out — Method

show_tx_out(auth::UserAuth; txid::String, n::Int64 = 1, include_mempool::Bool = true)

Returns details about an unspent transaction output.

BitcoinRPC.show_tx_out_setinfo — Method

Returns statistics about the unspent transaction output set.

BitcoinRPC.verify_chain — Method

verify_chain(auth::UserAuth; checklevel::Int64 = 1, nblocks::Int64 = 10)

Verifies blockchain database.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Optional keywords

checklevel::Int64 : Control the thoroughness of block verification. Each level includes the checks of the previous levels. Default is level 1.

                  - level 0 reads the blocks from disk
                  - level 1 verifies block validity
                  - level 2 verifies undo data
                  - level 3 checks disconnection of tip blocks
                  - level 4 tries to reconnect the blocks

nblocks::Int64 : number of blocks to check, default is set to 10

Example

julia> verify_chain(auth, nblocks = 100)
true

Mining RPCs

BitcoinRPC.show_mining_info — Method

show_mining_info(auth::UserAuth)

Returns a json object containing mining-related information.

BitcoinRPC.show_network_hashps — Method

show_network_hashps(auth::UserAuth; nblocks::Int64 = -1, height::Int64 = -1)

Returns the estimated network hashes per second based on the selected number of blocks.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

Optional keywords

nblocks::Int64 : Size of the window in number of blocks, used for estimation of the network hashes per second. Default is set to -1, which specifies since last difficulty change.

height::Int64 : Estimate the network speed at the time when a certain block was found.

Example

julia> show_network_hashps(auth, nblocks = 144)
2.435873902493952e20

julia> show_network_hashps(auth, height = 144)
900742.2065335643

Perform on-chain analytics

BitcoinRPC.collect_block_stats — Method

collect_block_stats(auth::UserAuth, block_start::Int64, block_end::Int64)

Collect statistics by iterating over a range of blocks.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

block_start::Int64 : Starting block height

block_end::Int64 : Ending block height

Example

julia> collect_block_stats(auth, 700_000, 700_005)
6×28 DataFrame
 Row │ avgfee      avgfeerate  avgtxsize  blockhash                          height  ins    maxfee      maxfeerate  maxtxsize  medianfe ⋯
     │ Float64     Float64     Int64      String                             Int64   Int64  Float64     Float64     Int64      Float64  ⋯
─────┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ 0.00012069      1.5e-7       1000  0000000000000000000590fc0f3eba19…  700000   6342  0.014          3.58e-6      86228   1.512e- ⋯
   2 │ 2.078e-5        2.0e-8        956  00000000000000000002f39baabb00ff…  700001   2435  0.00153        2.51e-6      46044   3.36e-6
   3 │ 4.082e-5        8.0e-8        627  00000000000000000001993b6b5e4e3d…  700002    729  0.00059549     3.16e-6      22613   1.363e-
   4 │ 0.00025372      3.5e-7        981  000000000000000000055646f272b32b…  700003    716  0.025465       2.02e-6      35443   1.486e-
   5 │ 5.484e-5        1.1e-7        831  0000000000000000000e360e05cb9d7b…  700004   1174  0.0028332      4.77e-6      53046   1.323e- ⋯

BitcoinRPC.collect_network_stats — Method

collect_network_stats(auth::UserAuth, block_start::Int64, block_end::Int64)

Collect block and network statistics by iterating over a range of blocks.

Arguments

auth::UserAuth : User credentials, e.g. auth = UserAuth("username", "password", port)

block_start::Int64 : Starting block height

block_end::Int64 : Ending block height

Example

julia> collect_network_stats(auth, 700_000, 700_100)
101×4 DataFrame
 Row │ height  time                 network_hash  difficulty 
     │ Int64   DateTime             Float64       Float64    
─────┼───────────────────────────────────────────────────────
   1 │ 700000  2021-09-11T04:14:32    1.29538e20  1.80963e13
   2 │ 700001  2021-09-11T04:15:02    1.29813e20  1.81346e13
   3 │ 700002  2021-09-11T04:17:07    1.30042e20  1.81666e13
   4 │ 700003  2021-09-11T04:17:57    1.30306e20  1.82036e13
   5 │ 700004  2021-09-11T04:20:45    1.30515e20  1.82327e13